updated tab completion
